refactor(hs): create utils/types.ts
