+++ packages/headless-styles/src/components/types.ts
@@ -31,7 +31,6 @@ export interface InputFieldOptions extends FieldOptions {
 
 export type CheckboxDirection = 'row' | 'col'
 export type IconSize = 's' | 'm' | 'l'
-export type Tech = 'svelte' | ''
 export type Position =
   | 'topStart'
   | 'top'
@@ -45,3 +44,5 @@ export type Position =
   | 'leftStart'
   | 'left'
   | 'leftEnd'
+export type Sentiment = 'default' | 'info' | 'success' | 'warning' | 'danger'
+export type Tech = 'svelte' | undefined
